[#145998989] ubi: Add backwards compatibility for oe-classic u-boot  … (#3)
* README: move to Microchip in this file

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>

* devmem2: tmp add devmem2 recipe in meta-atmel

devmem2 recipe is updated in mainline openembedded but not yet backported
in morty. Add it here as a workaround.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>

* [#145998989] ubi: Add backwards compatibility for oe-classic u-boot

---
The oe-classic u-boot version tries to mount a volume named 'nedap9g45-rootfs'. To be able to support a mixed setup (oe-classic u-boot and oe-core rootfs) the oe-core rootfs ubi volumename must match the name used in oe-classic.